ER-6 Страница управления занятиями

+++ src/main/java/ru/innopolis/stc9/earth_stc9/controllers/LessonsController.java
@@ -13,8 +13,11 @@ import ru.innopolis.stc9.earth_stc9.pojo.Lesson;
 import ru.innopolis.stc9.earth_stc9.pojo.Subject;
 import ru.innopolis.stc9.earth_stc9.pojo.User;
 import ru.innopolis.stc9.earth_stc9.services.ILessonService;
+import ru.innopolis.stc9.earth_stc9.services.IUsersService;
 
 import javax.servlet.http.HttpServletRequest;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -26,19 +29,21 @@ public class LessonsController {
     private static final Logger logger = Logger.getLogger(LessonsController.class);
     @Autowired
     private ILessonService serviceLesson;
+    @Autowired
+    private IUsersService serviceUser;
 
     @RequestMapping(value = "/lessons", method = RequestMethod.GET)
     public String doGet(HttpServletRequest req, Model model) {
         String login = ((String) req.getSession().getAttribute("login"));
         setLessons(login, model);
-        setFields(model);
+        setFields(login, model);
         return "lessons";
     }
 
     private void setLessons(String login, Model model) {
         List<Lesson> lessons = null;
         if (login != null && !login.isEmpty()) {
-            User userByLogin = serviceLesson.getUserByLogin(login);
+            User userByLogin = serviceUser.getUserByLogin(login);
             switch (userByLogin.getRole().getId()) {
                 case Roles.ADMIN_ROLE_ID:
                     lessons = serviceLesson.getLessonsLast(100);
@@ -56,33 +61,50 @@ public class LessonsController {
         model.addAttribute("lessons", lessons);
     }
 
-    private void setFields(Model model) {
+    private void setFields(String login, Model model) {
 
-        List<User> teachers = serviceLesson.getTeachers();
+        List<User> teachers = new ArrayList<>();
         List<Subject> subjects = serviceLesson.getSubjects();
         List<Group> groups = serviceLesson.getGroups();
+
+        if (login != null && !login.isEmpty()) {
+            User userByLogin = serviceUser.getUserByLogin(login);
+            switch (userByLogin.getRole().getId()) {
+                case Roles.ADMIN_ROLE_ID:
+                    teachers = serviceLesson.getTeachers();
+                    break;
+                case Roles.TEACHER_ROLE_ID:
+                    teachers.add(serviceUser.getUserByLogin(login));
+                    break;
+                default:
+                    break;
+            }
+        }
 
         model.addAttribute("subjects", subjects);
         model.addAttribute("teachers", teachers);
         model.addAttribute("groups", groups);
     }
 
-    @RequestMapping(value = "/lessonAdd", method = RequestMethod.POST)
+    @RequestMapping(value = "/lessons/add", method = RequestMethod.POST)
     public String addLesson(HttpServletRequest req, Model model,
                             @RequestAttribute String subject, @RequestAttribute String teacher,
-                            @RequestAttribute String theme, @RequestAttribute String group) {
+                            @RequestAttribute String theme, @RequestAttribute String group,
+                            @RequestAttribute String lessDate) {
         try {
             Lesson lesson = null;
             int subjectId = Integer.parseInt(subject);
             int groupId = Integer.parseInt(group);
+            SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
+            Date parsedDate = format.parse(lessDate);
 
             if (teacher == null || teacher.isEmpty()) {
                 String login = (String) req.getSession().getAttribute("login");
-                User userByLogin = serviceLesson.getUserByLogin(login);
-                lesson = new Lesson(theme, new Date(), subjectId, userByLogin.getId(), groupId);
+                User userByLogin = serviceUser.getUserByLogin(login);
+                lesson = new Lesson(theme, parsedDate, subjectId, userByLogin.getId(), groupId);
             } else {
                 int teacherId = Integer.parseInt(teacher);
-                lesson = new Lesson(theme, new Date(), subjectId, teacherId, groupId);
+                lesson = new Lesson(theme, parsedDate, subjectId, teacherId, groupId);
             }
             serviceLesson.createLesson(lesson);
         } catch (Exception e) {
@@ -91,15 +113,57 @@ public class LessonsController {
         return doGet(req, model);
     }
 
-    @RequestMapping(value = "/lessonDelete", method = RequestMethod.POST)
+    @RequestMapping(value = "/lessons/save", method = RequestMethod.POST)
+    public String saveLesson(HttpServletRequest req, Model model,
+                             @RequestAttribute String subject, @RequestAttribute String teacher,
+                             @RequestAttribute String theme, @RequestAttribute String group,
+                             @RequestAttribute String lessDate, @RequestAttribute String editLessonId) {
+        try {
+            int subjectId = Integer.parseInt(subject);
+            int groupId = Integer.parseInt(group);
+            int teacherId = Integer.parseInt(teacher);
+            SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
+            Date parsedDate = format.parse(lessDate);
+
+            Lesson lesson = new Lesson(Integer.parseInt(editLessonId), theme, parsedDate, subjectId, teacherId, groupId);
+            serviceLesson.changeLesson(lesson);
+        } catch (Exception e) {
+            logger.warn("Changed lesson saving", e);
+        }
+        return doGet(req, model);
+    }
+
+    @RequestMapping(value = "/lessons/delete", method = RequestMethod.POST)
     public String deleteLesson(HttpServletRequest req, Model model, @RequestAttribute String lessonId) {
         try {
             int idLesson = Integer.parseInt(lessonId);
-            serviceLesson.deleteLessonById(idLesson);
+            if (serviceLesson.existsInJournal(idLesson)) {
+                String message = "Невозможно удаление сведений о занятии, в журнале есть оценка";
+                model.addAttribute("message", message);
+            } else {
+                serviceLesson.deleteLessonById(idLesson);
+            }
         } catch (Exception e) {
             logger.warn("Delete lesson", e);
         }
         return doGet(req, model);
     }
 
+    @RequestMapping(value = "/lessons/change", method = RequestMethod.POST)
+    public String changeLesson(HttpServletRequest req, Model model,
+                               String lessonId) {
+        try {
+            Lesson editLesson = null;
+            editLesson = serviceLesson.getById(Integer.parseInt(lessonId));
+            if (serviceLesson.existsInJournal(Integer.parseInt(lessonId))) {
+                String message = "Невозможно редактирование сведений о занятии, в журнале есть оценка";
+                model.addAttribute("message", message);
+            } else {
+                model.addAttribute("editLesson", editLesson);
+            }
+        } catch (Exception e) {
+            logger.warn("Change lesson begin", e);
+        }
+        return doGet(req, model);
+    }
 }
